Fixing bugs

+++ decrypt.py
@@ -5,37 +5,48 @@ from cryptography.fernet import Fernet
 
 files = os.listdir()
 targets = []
-
-
-# Get encryption key
 key = 'key.key'
-with (open(key, 'rb')) as the_key:
-    encryption_key = the_key.read()
 
 
-# Get all files from the current directory
-for file in files:
+# Check if key exist in the current directory
+if os.path.exists(key):
 
-    file_extension = os.path.splitext(file)
+    # Get encryption key
+    with (open(key, 'rb')) as the_key:
+        encryption_key = the_key.read()
 
-    # Execlude Python scripts & key from decryption
-    if file.lower().endswith(('.py', '.key')):
-        continue
 
-    # Execlude directories from decryption
-    if os.path.isdir(os.path.abspath(file)):
-        continue
-    
-    targets.append(file)
+    # Get all files from the current directory
+    for file in files:
 
-    with open(file, 'rb') as the_file:
-        file_content = the_file.read()
+        file_extension = os.path.splitext(file)
 
-    decrypted_content = Fernet(encryption_key).decrypt(file_content)
-    with open(file, 'wb') as the_file:
-        the_file.write(decrypted_content)
+        # Execlude Python scripts & key from decryption
+        if file.lower().endswith(('.py', '.key')):
+            continue
 
-print('Congratulations ... Your Files Has Been Decrypted Successfully')
+        # Execlude directories from decryption
+        if os.path.isdir(os.path.abspath(file)):
+            continue
+        
+        targets.append(file)
+
+        with open(file, 'rb') as the_file:
+            file_content = the_file.read()
+
+        decrypted_content = Fernet(encryption_key).decrypt(file_content)
+        with open(file, 'wb') as the_file:
+            the_file.write(decrypted_content)
+
+    # Delete key after decryption
+    try:
+        os.remove(key)
+    except OSError as e:
+        print(f"Error in deleting key{e}")
+
+    print('Congratulations ... Your Files Has Been Decrypted Successfully')
+else:
+    print('Key is not exist!')
